Increase min distance on getNearestReefTag to guarantee a result
In testing today at Stuy Gabriel and I noticed that if we started
driveToClosestBranch too far away from an AprilTag, the whole robot would
reset. We saw on SmartDashboard that even though though the camera could
see and was recognizing a tag, closestTag was -1. Going through the code
I see that aprilTagFieldLayout.getTagPose returns an optional and when
you try to `get` on it when it's null it throws an exception.

I think the simplest fix is to increase min_distance to 30. Unless our
pose is incredibly messed up, which would be a more serious bug and
would preclude any driveToClosestBranch working, this should basically
guarantee that we always find a closest tag.