Refactored the elevator move method.
No more separate elevate/descend. I updated a number of commands to use
the new method (and to call setSpeed 0 instead of stop but that one's
debatable), but I'm not sure if they're used, I think they can just be
deleted.

I think I can also refactor the Elevator a bit more because it has a lot
of if statements that end up calling the same thing.