Some more refactors on elevator.
I removed some of the smartdashboard signals. In a way we already
display the information since we already display the "coral gap blocked"
sensor and we also already display the elevator position. It takes a bit
more effort to read if it's < 3 and read the other value, but we can add
a smartdashboard element back in. Overall if we can use fewer
smartdashboard values it can help to make our display less cluttered and
more intelligible.

Also important, we were using these functions for both disply and
functionality, and that's dangerous, because if you think "oh, we don't
need to display this" and you comment it out, well now not only are you
not displaying the thing but you've lost functionality as well.

So there's a lesson to be learned that dispalying something should have
no side effects.