Some more elevator refactor
One of the things we'll need to test is for the elevator to stop if
there's a coral blocking the gap in both joystick operation and
autonomous operation.

I'm not 100% sure if I did the right thing in removing the if statement
from ElevatorSetPosition, but I think I did. The only difference in the
code is that if the setpoint was intake height and there was a coral,
the existing code would not stop, and now it does...

If it's holding at intake height, maybe that's ok? We really need to set
brake mode on the elevator motors anyway.