Move some commands into Robot::teleopPeriodic
I've only moved the ones that aren't used in any autonomous commands,
but this does raise the question of how to do auto. In theory we could
make Commands within the auto section that set the motors, but this gets
clunky really fast.

Maybe we just configure the controller in teleopInit with the relevant
commands. We could in theory just do commands for auto, but it might be
nice to use the commands with the joystick as well, then if we change
something we can change it in one place.

Maybe we should do it in robotinit though, or even earlier, just because
if it takes more than 20ms to set the bindings, I don't want us going
overbudget in the switch from auto during a match. I don't see any
reason configuring buttons should take any more than fractions of a
microsecond, but better safe than sorry.